requirements.txt
apache-airflow
apache-airflow-providers-mysql
mysql-connector-python
pandas
requests
apache-superset
pyenv

